requirements.txt
GitPython==3.1.41
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
click==8.1.7